+++ frontend-react/src/AppRouter.tsx
@@ -6,7 +6,7 @@ import { RequireGate } from "./shared/RequireGate/RequireGate";
 import { SenderType } from "./utils/DataDashboardUtils";
 import { lazyRouteMarkdown } from "./utils/LazyRouteMarkdown";
 import { PERMISSIONS } from "./utils/UsefulTypes";
-const ReportTestingPage = lazy(() => import("./components/Admin/ReportTesting"));
+const ReportTestingPage = lazy(() => import("./components/Admin/MessageTesting"));
 /* Content Pages */
 const Home = lazy(lazyRouteMarkdown(() => import("./content/home/index.mdx")));
 const About = lazy(lazyRouteMarkdown(() => import("./content/about/index.mdx")));
